styling init with bottombar
