+++ components/src/modern/mod.rs
@@ -1,0 +1,1 @@
+pub mod bottombar;
