+++ components/src/normal/mod.rs
@@ -1,0 +1,1 @@
+pub mod bottombar;
